+++ tests/src/lib.rs
@@ -444,9 +444,11 @@ fn run_tests(cfg: TestConfig) -> Result<bool, io::Error> {
 
                 r#box::space_get_by_name,
                 r#box::space_get_by_name_cached,
+                r#box::space_cache_invalidated,
                 r#box::space_get_system,
                 r#box::index_get_by_name,
                 r#box::index_get_by_name_cached,
+                r#box::index_cache_invalidated,
                 r#box::insert,
                 r#box::replace,
                 r#box::delete,